Defined more interface as per `graphql-js` - Type - Selection - Value
